requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<3.15
django-cors-headers>=4.4,<4.5
channels>=4.0,<5.0
channels-redis>=4.1,<5.0
celery>=5.3,<5.4
redis>=4.5,<5.0
psycopg[binary]>=3.1,<3.2
python-dotenv>=1.0,<1.1
pandas>=2.1,<2.3
numpy>=1.26,<1.27
ccxt>=4.1,<4.2
